Deserialize frame's payload into Response
- Remove `From<Frame> for Response`.
- Return `Result<Start<R>, Error>` from `start_state(&self)`. This
  allows to return the `rfc003::Error` held inside the `Error` state
  which is currently the only state which can't produce the `Start`
  struct.
- Remove `get_swap` macro. There's no need for it since the types
  could already be extracted using the macro `with_swap_types`.
