+++ application/comit_node/src/http_api/rfc003/action.rs
@@ -10,7 +10,6 @@ use crate::{
                 ActionKind, Actions,
             },
             bitcoin, ethereum,
-            roles::{Alice, Bob},
             state_machine::StateMachineResponse,
             state_store::StateStore,
             Ledger, SecretSource,
@@ -538,14 +537,13 @@ fn handle_get<T: MetadataStore<SwapId>, S: StateStore<SwapId>>(
     let metadata = metadata_store
         .get(id)?
         .ok_or_else(problem::swap_not_found)?;
-    get_swap!(
+    with_swap_types!(
         &metadata,
-        state_store,
-        id,
-        state,
         (|| {
-            let state =
-                state.ok_or_else(|| HttpApiProblem::with_title_and_type_from_status(500))?;
+            trace!("Fetched metadata of swap with id {}: {:?}", id, metadata);
+            let state = state_store
+                .get::<Role>(id)?
+                .ok_or_else(problem::state_store)?;
             trace!("Retrieved state for {}: {:?}", id, state);
 
             state
