+++ application/comit_node/src/http_api/rfc003/mod.rs
@@ -1,4 +1,4 @@
 #[macro_use]
-mod get_state;
+mod with_swap_types;
 pub mod action;
 pub mod swap;
